:recycle: split up sources module to accomodate for future additions
